Crate governance_os_pallet_coin_voting

This pallet can be used to deploy a coin voting system. It supports direct, simple coin voting as well as the possibility to configure it to support quadratic voting.

Structs

Locks
Keeps track of locks set on user's balances and to which proposal they were linked to.

Module

Proposals
Proposals actively opened and linked to this voting implementation. Erased when closed or vetoed.

VoteData

VotingParameters

Enums

Call
Dispatchable calls.

Error

VoteCountingStrategy
Various strategies we support to count votes.

Constants

COIN_VOTING_LOCK_ID

Traits

Trait